db_setup/requirements.txt
psycopg2-binary==2.9.10
python-dotenv==1.1.1